+++ declare/CHANGELOG.md
@@ -2,3 +2,9 @@
 
 - Initial release of `declare`.
 - Features: State<T>, ViewModel, Published<T>, StateBuilder, ViewModelBuilder.
+
+## 1.0.0 - 2025-06-06
+
+- Fixed notifyListeners issue
+- Added hot reload support for state preservation
+- Changed name of ViewModelBuilder -> DeclareView